requirements.txt
Django~=3.2.6
Pillow~=9.1.1